requirements.txt
bootstrap-flask~=2.2.0
flask~=2.2.2
flask-bootstrap~=3.3.7.1
flask-login~=0.6.2
flask-sqlalchemy~=3.0.2
flask-wtf~=1.0.1
email-validator~=1.3.0
python-dotenv~=0.21.0
